;kl;kl;klkklhjkadsasdasd

+++ NonSteamLaunchers.sh
@@ -2891,9 +2891,23 @@ fi
 
 
 
+# Get the command line arguments
+args=("$@")
 
+# Check if there are any command line arguments
+if [ ${#args[@]} -gt 0 ]; then
+    # If there are command line arguments, kill all instances of Steam
+    killall steam
 
-killall steam
+    # Wait for all instances of Steam to be closed
+    while pgrep steam > /dev/null; do sleep 1; done
+else
+    # If there are no command line arguments, detach the script from the Steam process
+    nohup sh -c 'sleep 10; /usr/bin/steam' &
+
+    # Wait for the steam process to exit
+    while pgrep steam > /dev/null; do sleep 1; done
+fi
